feat: new details to be retrieved from users
