+++ package.json
@@ -11,6 +11,7 @@
   "dependencies": {
     "@expo/vector-icons": "^14.1.0",
     "@react-native-async-storage/async-storage": "2.1.2",
+    "@react-native-picker/picker": "^2.11.0",
     "@react-navigation/bottom-tabs": "^7.3.10",
     "@react-navigation/native": "^7.1.6",
     "@stripe/stripe-react-native": "^0.47.0",
@@ -33,6 +34,7 @@
     "react-dom": "^19.0.0",
     "react-native": "0.79.2",
     "react-native-chart-kit": "^6.12.0",
+    "react-native-element-dropdown": "^2.12.4",
     "react-native-gesture-handler": "~2.24.0",
     "react-native-gifted-charts": "^1.4.61",
     "react-native-paper": "^5.14.4",
